klte-common: Disable nonblocking_ffs to avoid recent crashes with adbd Recent changes to adbd have enabled nonblocking_ffs by default.
These have seeminly uncovered some bad memory corruption in the
kernel ffs gadget driver on every kernel I've tried so far.

Unfortunately I've not been able to diagnose what is going wrong,
so for now, lets disable the nonblocking_ffs.

We still need to root cause the issue as its a problem even with
the mainline kernel on HiKey, so its likely a generic kernel issue.

Change-Id: I81ac10edf5ddf9040852a4cb1f0f07c4ca16ef67